REFACTOR: Move pages, components, lib and styles into src/
