+++ src/pages/_app.js
@@ -1,4 +1,4 @@
-import "../node_modules/uikit/dist/css/uikit.min.css";
+import "../../node_modules/uikit/dist/css/uikit.min.css";
 import "../styles/style.css";
 
 function App({ Component, pageProps }) {
