+++ src/pages/resources.js
